+++ poem-openapi/CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+# [1.0.43] 2021-12-07
+
+- Change the schema type of enum to `string` [#121](https://github.com/poem-web/poem/issues/121)
+
 # [1.0.42] 2021-12-07
 
 - Implement `Type` for `&[T]`, `&T` and `[T; const N: usize]`.